Add files via upload
This is a machine learning project aimed at building a lead scoring system for businesses. The project involves training a model on a dataset of leads with their corresponding features and status, and predicting a lead score ranging from 0-100 based on the input set of features. The project also involves dropping the leads with a status other than "WON" or "LOST" during training and treating all columns as categorical. The resulting model can help businesses prioritize leads and increase their conversion rate by allowing the sales team to focus on potential customers that are more likely to convert into actual bookings.